split CustomSearch to components
